refactor: 로또 번호 형식 검증 validator 제거
